requirements.txt
langchain>=0.3.0,<0.4
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
python-docx>=1.1.0
sentence-transformers>=2.2.0
torch>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
fastmcp>=2.0.0
